adk-gemini3/video_monitor_agent/requirements.txt
google-adk>=0.5.0
google-genai
